requirements.txt
# SynapseGrid Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
aioredis==2.0.1
asyncpg==0.29.0
grpcio==1.59.3
grpcio-tools==1.59.3
pydantic==2.5.0
python-multipart==0.0.6
prometheus-client==0.19.0
pyjwt==2.8.0
websockets==12.0

# AI/ML Dependencies (for Mac M2 node)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.21.0
numpy>=1.21.0
pillow>=9.0.0
onnxruntime>=1.15.0

# System monitoring
psutil>=5.9.0

# HTTP client
aiohttp>=3.8.0
aiofiles>=23.0.0
requests>=2.28.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
